requirements.txt
alembic==1.9.2
autopep8==2.0.1
certifi==2022.12.7
click==8.1.3
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy==0.991
mypy-extensions==0.4.3
pipenv==2023.2.4
platformdirs==3.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.0
SQLAlchemy-Utils==0.39.0
tomli==2.0.1
typing_extensions==4.4.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.2.2
WTForms==3.0.1
